feat: Add new network interface to allow host->container communication
By design, the `macvlan` network type does allow communication between
the host and devices on the network. This means that the host cannot
talk to the containers on the services or vice versa. This is now an
issue as I want to retrieve metrics and other observability data from
the services.

In order to allow this we need to create a new `macvlan` network in
bridge mode, add an interface, and then assign a static route from the
host to this interface.

Big ups:
 * Barebones commands: https://stackoverflow.com/a/57038734/147284
 * Explanation why this doesn't work: https://stackoverflow.com/a/44049431/147284
 * Skeleton NixOS config: https://github.com/NixOS/nixpkgs/blob/62f746b630d5189b2eedc19688ddcfb8021c1c71/nixos/tests/containers-macvlans.nix