feat: Caddy + automatic TLS certificates now works!
I had to build and run a custom Docker container
(https://github.com/richid/caddy-docker-proxy-r53-dns) to get the R53
ACME challenge code working correctly...but it does! And hopefully will
never need to mess with this again.

This relies on the following snippet in the Caddyfile:

*.fatsch.us, fatsch.us {
    tls {
        dns route53 {
            max_retries 10
            access_key_id <access_key>
            secret_access_key <secret_key>
            region "us-east-1"
            wait_for_propagation true
            hosted_zone_id <zone_id>
        }
    }
}

I've moved the Audiobookshelf service over to the new domain and it's
working great. Will move the others in the future.