feat: Enable tailscale
The current configuration starts the daemon but doesn't perform any of
the remaining configuration. Specifically, you need to run
`tailscale up --auth-key=<auth-key>` to authenticate to the control
servers. Then, to enable subnet routing, you'll need to run
`tailscale set --advertise-routes "192.168.20.0/24"`. There is a way to
do this with a custom one-shot systemd unit but that is for another day.

This works though! The current network configuration has public DNS
entries returning private IPs to the Caddy reverse proxy. Once the
server is advertising these routes you can connect over the public 'net
when tailscale is enabled.